Refactored background layer & fixed compilation issues
Finished refactoring out the BackgroundImage component from the base
Battlemap component by introducing a BackgroundLayer component to manage
the state and position of all background images contained within it.
This makes the management of the various layers within the Battlemap far
easier to manage by establishing each layer to be a self-contained
component with separate state.

Further refactored project folder structure to align with the new
standard. While the changes are not complete, this does make component
hierarchy somewhat easier to follow (theoretically).

Fixed all compilation issues in the project:

  - Image SVG element within the BackgroundImage component had an
  incorrect attribute. xlink:href was changed to the correct xlinkHref.

  - Fixed several issues with the wrong variable being called/accessed.