Start redux integration to replace internal store
Redux reducers have been added & barely plugged in order to start the
transfer from mainly relying on React's store system to using Redux.
This change will allow the relevant state changes to propagate to the
other parts of the application that rely on it, and will make it far
easier to serialize/deserialize the application when server
communication is added.